+++ doc/source/changelog.rst
@@ -3,6 +3,19 @@
 ChangeLog
 =========
 
+Version 4.4.2
+-------------
+
+This version of pyXSIM contains a bugfix for simulations where particle-based
+emission sources have no intrinsic widths (neither cell sizes or smoothing
+lengths).
+
+Version 4.4.1
+-------------
+
+This version of pyXSIM contains a bugfix, where supplying an array of floats
+for abundance tables in thermal source models was not working correctly.
+
 Version 4.4.0
 -------------